+++ src/commands/index.ts
@@ -2,3 +2,4 @@ export * from './tokenCommands';
 export * from './secretCommands';
 export * from './variableCommands';
 export * from './runCommands';
+export * from './workflowCommands';